[storage] half done storage stuff
