+++ packages/shelter/src/plugins.tsx
@@ -1,5 +1,5 @@
-import { isInited, signalOf, solidMutWithSignal, storage, waitInit } from "./storage";
-import { Component } from "solid-js";
+import { isInited, signalOf, solidMutWithSignal, idbStore, waitInit } from "./storage";
+import { Component, createSignal, JSX } from "solid-js";
 import { createMutable } from "solid-js/store";
 import { createScopedApi, log } from "./util";
 import { ModalBody, ModalHeader, ModalRoot, openModal } from "@uwu/shelter-ui";
@@ -23,8 +23,8 @@ export type EvaledPlugin = {
   scopedDispose(): void;
 };
 
-const internalData = storage<StoredPlugin>("plugins-internal");
-const pluginStorages = storage("plugins-data");
+const internalData = idbStore<StoredPlugin>("plugins-internal");
+const pluginStorages = idbStore("plugins-data");
 const [internalLoaded, loadedPlugins] = solidMutWithSignal(createMutable({} as Record<string, EvaledPlugin>));
 
 export const installedPlugins = signalOf(internalData);
